+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Parameter title for er and flow diagrams.
+
 ## [0.3.0] - 2023-03-15
 
 ### Added